pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "brain-mcp"
version = "0.1.0"
description = "Turn your AI conversations into a searchable second brain with cognitive prosthetic tools"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "Mordechai Potash"}]
keywords = ["mcp", "brain", "ai", "conversations", "search", "cognitive", "prosthetic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "mcp>=1.0",
    "lancedb>=0.20",
    "duckdb>=1.0",
    "pandas>=2.0",
    "pyarrow>=15.0",
    "sentence-transformers>=2.0",
    "pyyaml",
    "rich>=13.0",
]

[project.optional-dependencies]
summarize = ["anthropic>=0.40", "openai>=1.0"]
dev = ["pytest>=7.0", "build", "twine"]

[project.scripts]
brain-mcp = "brain_mcp.cli:main"

[project.urls]
Homepage = "https://github.com/mordechaipotash/brain-mcp"
Documentation = "https://brainmcp.dev"
Repository = "https://github.com/mordechaipotash/brain-mcp"

[tool.setuptools.packages.find]
include = ["brain_mcp*"]
